show snapshot list
