+++ frontend/src/features/collections/collection-thumbnail.ts
