+++ frontend/src/features/collections/index.ts
@@ -2,8 +2,9 @@ import("./collections-add");
 import("./collections-grid");
 import("./collection-items-dialog");
 import("./collection-metadata-dialog");
+import("./collection-start-page-dialog");
 import("./collection-workflow-list");
 import("./select-collection-access");
-import("./collection-start-page-dialog");
+import("./select-collection-start-page");
 import("./share-collection");
-import("./thumbnail");
+import("./collection-thumbnail");
